frontend/requirements.txt
# Chainlit for the chat interface
chainlit==0.7.501

# HTTP client for backend communication (compatible with chainlit)
httpx>=0.23.0,<0.25.0

# Additional utilities
python-dotenv==1.0.0
pydantic>=1.10.0,<2.0.0 